requirements.txt
streamlit
pymupdf
python-dotenv
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
